+++ src/store/features/buildGame/buildGameMiddleware.ts
@@ -51,7 +51,7 @@ const buildGameMiddleware: Middleware<Dispatch, RootState> =
       const dispatch = store.dispatch.bind(store);
       await API.app.deleteBuildCache();
       dispatch(consoleActions.clearConsole());
-      dispatch(consoleActions.stdOut("Cleared GB Studio caches"));
+      dispatch(consoleActions.stdOut({ text: "Cleared GB Studio caches" }));
     } else if (actions.ejectEngine.match(action)) {
       API.project.ejectEngine();
     } else if (actions.exportProject.match(action)) {